Adding logic to support primary and secondary server connections. Also fixing class static var issues.
Basically, this changes the summoning logic. Instead of a summon disconnecting and reconnecting to a new server, the plugin and service now support a primary and secondary connection.

The primary connection uses the starport.octoeverywhere.com hostname, which should dynamically resolve to the best perf server. This connection is used for the majority of communication to the service.

However, if a user shares a shared connection with someone that's outside of their region, the shared connection will resolve to their local OctoEverywhere server. The old summon logic would disconnect and reconnect to the other server. The problem with that is if the first user was also on, it would break their connection.

So this new logic allows a new concept called secondary servers. These connections can be made to any OctoEverwhere server and are only supported for a short time. This allows the summon to service the user in a different region, but for the printer to also stay active in the current region.