Welcome to OctoEverywhere 1.0!
This commit is the first in a month or two long process to upgrade off the old OctoEverywhere protocol (now called v0) to our new protocol, v1! V0 served us well, but we have learned a lot since the first day we put pen to paper for OctoEverywhere. V1 brings a ton of great enhancements right now, and opens the door for so much more to come in the future!